# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LEAD CONTROL API",
        "version": "1.0"
    },
    "paths": {
        "/api/admin/registration": {
            "post": {
                "tags": [
                    "Admin API"
                ],
                "summary": "Register admin/operator",
                "operationId": "38178cb84f8cc67dc37d09fac0f6e122",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "Admin token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "example": "operator"
                                    },
                                    "login": {
                                        "type": "string",
                                        "example": "998994630613"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registration admin and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": null
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Added successfully!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/login": {
            "post": {
                "tags": [
                    "Admin API"
                ],
                "summary": "Login admin/operator",
                "operationId": "73502e033d895d7a98710fdedb3c2b3b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "login": {
                                        "type": "string",
                                        "example": "998994630613"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check admin **login** & **password** and to login.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 2,
                                                "admin_socket_id": null,
                                                "login": "998994630613",
                                                "password": "$2y$10$veIbPHbXbOknIz1qHRhNze3mIOI2DmX4jcJWupT96tA3vK8XAIZ5q",
                                                "token": "XsfiFH57c8AdWSC9EzquHWAkPPKnQS",
                                                "role": "operator",
                                                "status": "active",
                                                "created_at": "2024-01-12T09:13:46.000000Z",
                                                "updated_at": "2024-01-12T09:17:38.000000Z"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Welcome!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/get": {
            "get": {
                "tags": [
                    "Admin API"
                ],
                "summary": "Get admin/operator",
                "operationId": "593fa647e2d566b0e27fba408d999ded",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "Admin token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check admin **token** and get this admin.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 2,
                                                "admin_socket_id": null,
                                                "login": "998994630613",
                                                "password": "$2y$10$veIbPHbXbOknIz1qHRhNze3mIOI2DmX4jcJWupT96tA3vK8XAIZ5q",
                                                "token": "XsfiFH57c8AdWSC9EzquHWAkPPKnQS",
                                                "role": "operator",
                                                "status": "active",
                                                "created_at": "2024-01-12T09:13:46.000000Z",
                                                "updated_at": "2024-01-12T09:17:38.000000Z"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": ""
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/detail/{admin_id}": {
            "get": {
                "tags": [
                    "Admin API"
                ],
                "summary": "Get detail admin/operator",
                "operationId": "981a81167f684a7ab1420f5c86fb3dd7",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "Admin token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "admin_id",
                        "in": "path",
                        "description": "Paste admin id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check admin **token** and get detail admin.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 2,
                                                "admin_socket_id": null,
                                                "login": "998994630613",
                                                "password": "$2y$10$veIbPHbXbOknIz1qHRhNze3mIOI2DmX4jcJWupT96tA3vK8XAIZ5q",
                                                "token": "XsfiFH57c8AdWSC9EzquHWAkPPKnQS",
                                                "role": "operator",
                                                "status": "active",
                                                "created_at": "2024-01-12T09:13:46.000000Z",
                                                "updated_at": "2024-01-12T09:17:38.000000Z"
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": ""
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}